# openapi.yml
openapi: 3.1.0
info:
  title: gosearch
  version: 1.0.0

servers:
  - url: http://localhost:8080/api/v1

paths:
  /web:
    get:
      summary: Web search
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: engine
          in: query
          schema:
            type: string
            enum: [ddg, bing, brave]
            default: ddg
        - name: pages
          in: query
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WebResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "502":
          $ref: "#/components/responses/BadGateway"

  /software:
    get:
      summary: Software / repository search
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: source
          in: query
          schema:
            type: string
            enum: [github, gitlab, sourceforge]
            default: github
        - name: pages
          in: query
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: Repository results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Repository"
        "400":
          $ref: "#/components/responses/BadRequest"
        "502":
          $ref: "#/components/responses/BadGateway"

  /academic:
    get:
      summary: Academic paper search
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: source
          in: query
          schema:
            type: string
            enum: [openalex, nasa]
            default: openalex
        - name: pages
          in: query
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: Academic paper results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Paper"
        "400":
          $ref: "#/components/responses/BadRequest"
        "502":
          $ref: "#/components/responses/BadGateway"

  /vuln:
    get:
      summary: Vulnerability search
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: source
          in: query
          schema:
            type: string
            enum: [nvd, exploitdb]
            default: nvd
        - name: pages
          in: query
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: Vulnerability results
          content:
            application/json:
              schema:
                type: array
                items:
                  oneOf:
                    - $ref: "#/components/schemas/CVE"
                    - $ref: "#/components/schemas/Exploit"
        "400":
          $ref: "#/components/responses/BadRequest"
        "502":
          $ref: "#/components/responses/BadGateway"

  /apps:
    get:
      summary: Application search
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: source
          in: query
          schema:
            type: string
            enum: [flathub, homebrew]
            default: flathub
        - name: pages
          in: query
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: Application results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/App"
        "400":
          $ref: "#/components/responses/BadRequest"
        "502":
          $ref: "#/components/responses/BadGateway"

  /ml:
    get:
      summary: LLM/ML model search
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: source
          in: query
          schema:
            type: string
            enum: [ollama, huggingface]
            default: ollama
        - name: pages
          in: query
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: Model results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Model"
        "400":
          $ref: "#/components/responses/BadRequest"
        "502":
          $ref: "#/components/responses/BadGateway"

  /games:
    get:
      summary: Game search
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: source
          in: query
          schema:
            type: string
            enum: [steam, itchio, gog]
            default: steam
        - name: pages
          in: query
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: Game results
          content:
            application/json:
              schema:
                type: array
                items:
                  oneOf:
                    - $ref: "#/components/schemas/Game"
                    - $ref: "#/components/schemas/ItchGame"
                    - $ref: "#/components/schemas/GogGame"
        "400":
          $ref: "#/components/responses/BadRequest"
        "502":
          $ref: "#/components/responses/BadGateway"

components:
  schemas:
    App:
      type: object
      properties:
        AppID:
          type: string
        Name:
          type: string
        Summary:
          type: string
        Developer:
          type: string
        License:
          type: string
        Icon:
          type: string
          format: uri
        URL:
          type: string
          format: uri
        UpdatedAt:
          type: integer
          description: Unix timestamp

    WebResult:
      type: object
      properties:
        Title:
          type: string
        URL:
          type: string
          format: uri
        Snippet:
          type: string

    Repository:
      type: object
      properties:
        Name:
          type: string
        URL:
          type: string
          format: uri
        Description:
          type: string
        Stars:
          type: integer
        Language:
          type: string
        UpdatedAt:
          type: string
          format: date-time

    Error:
      type: object
      properties:
        error:
          type: string

    Paper:
      type: object
      properties:
        Title:
          type: string
        URL:
          type: string
          format: uri
        Authors:
          type: string
        Abstract:
          type: string
        Type:
          type: string

    CVE:
      type: object
      properties:
        ID:
          type: string
        Description:
          type: string
        Published:
          type: string
          format: date-time
        Severity:
          type: string
          enum: [LOW, MEDIUM, HIGH, CRITICAL]
        Score:
          type: number
        URL:
          type: string
          format: uri

    Exploit:
      type: object
      properties:
        ID:
          type: string
        Title:
          type: string
        Type:
          type: string
        Platform:
          type: string
        Author:
          type: string
        Published:
          type: string
          format: date
        CVEs:
          type: array
          items:
            type: string
        URL:
          type: string
          format: uri

    Model:
      type: object
      properties:
        name:
          type: string
        url:
          type: string
          format: uri
        description:
          type: string
        capabilities:
          type: array
          items:
            type: string
        pulls:
          type: string
        tags:
          type: string
        size:
          type: string
        updated:
          type: string

    Game:
      type: object
      properties:
        AppID:
          type: string
        Title:
          type: string
        URL:
          type: string
          format: uri
        ImageURL:
          type: string
          format: uri
        ReleaseDate:
          type: string
        Price:
          type: string
        OriginalPrice:
          type: string
        DiscountPercent:
          type: string
        ReviewSummary:
          type: string
        ReviewClass:
          type: string
        Platforms:
          type: array
          items:
            type: string

    ItchGame:
      type: object
      properties:
        GameID:
          type: string
        Title:
          type: string
        URL:
          type: string
          format: uri
        Author:
          type: string
        AuthorURL:
          type: string
          format: uri
        Description:
          type: string
        Genre:
          type: string
        ThumbnailURL:
          type: string
          format: uri
        Platforms:
          type: object
          properties:
            Windows:
              type: boolean
            MacOS:
              type: boolean
            Linux:
              type: boolean
            Web:
              type: boolean
            Android:
              type: boolean
        Rating:
          type: object
          nullable: true
          properties:
            Average:
              type: number
            Total:
              type: integer
            StarPercentage:
              type: number

    GogGame:
      type: object
      properties:
        ID:
          type: string
        Slug:
          type: string
        Title:
          type: string
        URL:
          type: string
          format: uri
        ImageURL:
          type: string
          format: uri
        ReleaseDate:
          type: string
        Price:
          type: string
        OriginalPrice:
          type: string
        DiscountPercent:
          type: string
        Developers:
          type: array
          items:
            type: string
        Publishers:
          type: array
          items:
            type: string
        Platforms:
          type: array
          items:
            type: string
        Genres:
          type: array
          items:
            type: string
        Tags:
          type: array
          items:
            type: string

  responses:
    BadRequest:
      description: Missing required parameter
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    BadGateway:
      description: Upstream source error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"